Moves currentdate function to a helper file
