+++ src/pages/events/index.js
@@ -3,6 +3,7 @@ import useSiteMetadata from "../../hooks/use-site-metadata";
 import Seo from "../../components/seo";
 import Layout from "../../components/Layout";
 import EventsRoll from "../../components/EventsRoll";
+import { getCurrentDate } from "../../utils/helpers";
 // import FeaturedProjectsTiles from "../../components/FeaturedProjectsTiles";
 import Donate from "../../components/Donate";
 // import HeroImage from "../../components/HeroImage";
@@ -19,21 +20,6 @@ const Section = styled.section`
   width: 100%;
 `;
 
-/**
- * Returns the current date in YYYY-MM-DD format
- */
-function getCurrentDate() {
-  const d = new Date();
-  let month = (d.getMonth() + 1).toString();
-  if (month.length < 2) {
-    month = `0${month}`;
-  }
-  let day = d.getDate().toString();
-  if (day.length < 2) {
-    day = `0${day}`;
-  }
-  return `${d.getFullYear()}-${month}-${day}`;
-}
 var currentDate = getCurrentDate();
 
 const EventsIndexPage = () => {
